+++ tests/remote_server/typescript-server/instrumentation.ts
@@ -1,6 +1,6 @@
 import { NodeSDK } from "@opentelemetry/sdk-node";
 import { LangfuseSpanProcessor } from "@langfuse/otel";
-import "./env";
+import "./env.js";
 
 const sdk = new NodeSDK({
   spanProcessors: [
