+++ tests/remote_server/typescript-server/tsconfig.json
